Controllers added

+++ src/main/java/ru/javarush/lukyanov/dao/TaskDAO.java
@@ -10,11 +10,10 @@ import ru.javarush.lukyanov.enums.Status;
 
 @Repository
 @Transactional(readOnly = true)
-public interface TaskDAO extends JpaRepository <Task, Long> {
+public interface TaskDAO extends JpaRepository<Task, Integer> {
     @Transactional
     @Modifying
     @Query("update Task t set t.description = ?1, t.status = ?2 where t.id = ?3")
     int updateTask(String description, Status status, Integer id);
-
 
 }
